+++ tests/publisher/test_main.py
@@ -50,8 +50,7 @@ def test_publisher_items_without_actors(mocked_backend: MagicMock) -> None:
                 "MergedVariable",
                 "MergedVariableGroup",
             ],
-            referenced_identifier=None,
-            reference_field=None,
+            reference_filters=None,
         ),
     ]
 
@@ -65,15 +64,13 @@ def test_publisher_persons(mocked_backend: MagicMock) -> None:
             publishing_target="invenio",
             query_string=None,
             entity_type=["MergedPerson"],
-            referenced_identifier=None,
-            reference_field=None,
+            reference_filters=None,
         ),
         call(
             publishing_target="invenio",
             query_string=None,
             entity_type=["MergedConsent"],
-            referenced_identifier=None,
-            reference_field=None,
+            reference_filters=None,
         ),
     ]
 
@@ -87,8 +84,7 @@ def test_publisher_contact_points_and_units(mocked_backend: MagicMock) -> None:
             publishing_target="invenio",
             query_string=None,
             entity_type=["MergedContactPoint", "MergedOrganizationalUnit"],
-            referenced_identifier=None,
-            reference_field=None,
+            reference_filters=None,
         )
     ]